refactor: move `getEpisodesBySeason` into `utils/transistor.server.ts`
